Enhancement move transform and other registration to separate files
