+++ client/src/boot/index.js
@@ -1,45 +1,15 @@
-import i18n from 'i18n';
-import { combineReducers } from 'redux';
 import Config from 'lib/Config';
 import reactRouteRegister from 'lib/ReactRouteRegister';
-import galleryReducer from 'state/gallery/GalleryReducer';
-import queuedFilesReducer from 'state/queuedFiles/QueuedFilesReducer';
 import AssetAdminRouter from 'containers/AssetAdmin/AssetAdminRouter';
-import uploadFieldReducer from 'state/uploadField/UploadFieldReducer';
-import previewFieldReducer from 'state/previewField/PreviewFieldReducer';
-import imageLoadReducer from 'state/imageLoad/ImageLoadReducer';
-import Injector from 'lib/Injector';
-import UploadField from 'components/UploadField/UploadField';
-import PreviewImageField from 'components/PreviewImageField/PreviewImageField';
-import ProportionConstraintField from 'components/ProportionConstraintField/ProportionConstraintField';
-import HistoryList from 'containers/HistoryList/HistoryList';
+import applyTransform from 'boot/applyTransform';
+import registerReducers from 'boot/registerReducers';
+import registerComponents from 'boot/registerComponents';
 
 document.addEventListener('DOMContentLoaded', () => {
-  Injector.component.register('UploadField', UploadField);
-  Injector.component.register('PreviewImageField', PreviewImageField);
-  Injector.component.register('HistoryList', HistoryList);
-  Injector.component.register('ProportionConstraintField', ProportionConstraintField);
+  registerComponents();
 
-  Injector.transform(
-    'insert-media-modal',
-    (updater) => {
-      updater.form.alterSchema('AssetAdmin.EditForm', form => {
-        const schema = form.getState();
-        const schemaId = schema.schema.id;
-        const overrides = schema.stateOverride && schema.stateOverride.fields;
-        const customTitle = (schemaId === 'Form_fileInsertForm' && overrides)
-          ? i18n._t('AssetAdmin.UPDATE_FILE', 'Update file')
-          : i18n._t('AssetAdmin.INSERT_FILE', 'Insert file');
+  applyTransform();
 
-        form.mutateField('action_insert', (field) => ({
-          ...field,
-          title: customTitle || field.title,
-        }));
-
-        return form.getState();
-      });
-    }
-  );
   const sectionConfig = Config.getSection('SilverStripe\\AssetAdmin\\Controller\\AssetAdmin');
 
   reactRouteRegister.add({
@@ -62,11 +32,5 @@ document.addEventListener('DOMContentLoaded', () => {
     ],
   });
 
-  Injector.reducer.register('assetAdmin', combineReducers({
-    gallery: galleryReducer,
-    queuedFiles: queuedFilesReducer,
-    uploadField: uploadFieldReducer,
-    previewField: previewFieldReducer,
-    imageLoad: imageLoadReducer,
-  }));
+  registerReducers();
 });
